Update gsport.py
Added -I -E -C -P options.

When -I is provided, the content of that file will be used to only download those files.

When -E is provided, the content of that file will be used to skip those files.

When -C is provided, files will be validated by downloading them twice, if the file is not in the online checksum.md5 file. This improves the odds of correctly downloading the file.

When -P is provided, the downloaded files will be stored in that path.